+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQ9RQW9NX6TKMBP6ZJZ5H6NG"
+        "01KQA457T23HDEZSNX8Q12N4V0"
     )
     r"""The unique identifier of the data source"""