+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-28T10:04:02.081Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-28T13:23:33.901Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")